Refactorisation de L'attribut "valeur" en "moyenne" Changement de type int vers float pour toutes les variable (min,max,moyenne) Création des urls pour les requêtes au système de données Remplacement de l'échafaud XML par le système de données Conversion du mois (chiffre) en texte pour affichage Ajout de la fonction getStatistique pour fournir les données manquantes aux vues

+++ Captures/src/modele/Sommaire.java
@@ -1,37 +1,37 @@
 package modele;
 
 public class Sommaire {
-	private int moyenneJour;
-	private int moyenneMois;
-	private int moyenneAnnee;
+	private float moyenneJour;
+	private float moyenneMois;
+	private float moyenneAnnee;
 	
 	public Sommaire() {
 		super();
 	}
 	
-	public Sommaire(int moyenneJour, int moyenneMois, int moyenneAnnee) {
+	public Sommaire(float moyenneJour, float moyenneMois, float moyenneAnnee) {
 		super();
 		this.moyenneJour = moyenneJour;
 		this.moyenneMois = moyenneMois;
 		this.moyenneAnnee = moyenneAnnee;
 	}
 	
-	public int getMoyenneJour() {
+	public float getMoyenneJour() {
 		return moyenneJour;
 	}
-	public void setMoyenneJour(int moyenneJour) {
+	public void setMoyenneJour(float moyenneJour) {
 		this.moyenneJour = moyenneJour;
 	}
-	public int getMoyenneMois() {
+	public float getMoyenneMois() {
 		return moyenneMois;
 	}
-	public void setMoyenneMois(int moyenneMois) {
+	public void setMoyenneMois(float moyenneMois) {
 		this.moyenneMois = moyenneMois;
 	}
-	public int getMoyenneAnnee() {
+	public float getMoyenneAnnee() {
 		return moyenneAnnee;
 	}
-	public void setMoyenneAnnee(int moyenneAnnee) {
+	public void setMoyenneAnnee(float moyenneAnnee) {
 		this.moyenneAnnee = moyenneAnnee;
 	}
